+++ .gitignore
@@ -9,6 +9,5 @@ pnpm-debug.log*
 node_modules
 dist
 
-# i sometimes use this to inspect json files :)
-ignore.json
-ignore.txt
+# i sometimes use this to inspect files :)
+ignore.*
