+++ package.json
@@ -38,25 +38,25 @@
 		"vue": "^3.2.37"
 	},
 	"devDependencies": {
-		"@nabla/vite-plugin-eslint": "^1.4.0",
-		"@types/node": "18.0.0",
-		"@typescript-eslint/eslint-plugin": "^5.30.5",
-		"@typescript-eslint/parser": "^5.30.5",
-		"@vitejs/plugin-vue": "^2.3.3",
+		"@nabla/vite-plugin-eslint": "^1.4.1",
+		"@types/node": "18.7.13",
+		"@typescript-eslint/eslint-plugin": "^5.35.1",
+		"@typescript-eslint/parser": "^5.35.1",
+		"@vitejs/plugin-vue": "^3.0.3",
 		"@vue/eslint-config-typescript": "^11.0.0",
-		"autoprefixer": "^10.4.7",
-		"eslint": "^8.19.0",
+		"autoprefixer": "^10.4.8",
+		"eslint": "^8.23.0",
 		"eslint-config-prettier": "^8.5.0",
 		"eslint-plugin-prettier": "^4.2.1",
-		"eslint-plugin-vue": "^9.1.1",
+		"eslint-plugin-vue": "^9.4.0",
 		"husky": "^8.0.1",
-		"postcss": "^8.4.14",
+		"postcss": "^8.4.16",
 		"prettier": "^2.7.1",
-		"prettier-plugin-tailwindcss": "^0.1.11",
-		"tailwindcss": "^3.1.4",
-		"typescript": "^4.7.4",
-		"vite": "^2.9.13",
+		"prettier-plugin-tailwindcss": "^0.1.13",
+		"tailwindcss": "^3.1.8",
+		"typescript": "^4.8.2",
+		"vite": "^3.0.9",
 		"vue-eslint-parser": "^9.0.3",
-		"vue-tsc": "^0.38.2"
+		"vue-tsc": "^0.40.1"
 	}
 }
